rename getInitialData to fetchPageData
