+++ src/client.test.jsx
@@ -23,7 +23,7 @@ jest.mock('@loadable/component', () => {
   };
 });
 
-jest.mock('./app/routes/getInitialData/utils/getRouteProps');
+jest.mock('./app/routes/fetchPageData/utils/getRouteProps');
 
 const mockRootElement = <div />;
 document.getElementById = jest.fn().mockReturnValue(mockRootElement);
